requirements.txt
# Core dependencies for Provider Network Management System
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
requests>=2.28.0

# PDF generation and image handling
reportlab>=4.0.0
Pillow>=9.5.0

# Excel export capabilities
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Configuration file support
PyYAML>=6.0

# Database support
sqlite3  # Usually included with Python

# Optional: Enhanced data processing
networkx>=3.0.0
scipy>=1.10.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Jupyter support (for development)
jupyter>=1.0.0
ipykernel>=6.25.0

# Mock data generation (for testing)
faker>=19.0.0